refactor: 유틸리티 함수 등을 util 패키지로 옮김
