--- app/src/main/java/hsk/practice/myvoca/Constants.kt
@@ -1,16 +1,0 @@
-package hsk.practice.myvoca
-
-/**
- * Constant values which is used at the whole application.
- */
-object Constants {
-    const val EDIT_VOCA: String = "VOCABULARY"
-    const val EDIT_NEW_VOCA_OK = 11
-    const val EDIT_NEW_VOCA_CANCEL = 12
-
-    /* Used for Fragment/Activity <-> AddVocaActivity */
-    const val CALL_ADD_VOCA_ACTIVITY = 13
-    const val ADD_NEW_VOCA_OK = 14
-    const val ADD_NEW_VOCA_CANCEL = 15
-
-}
