+++ app/src/main/java/hsk/practice/myvoca/ui/screens/allword/AllWordViewModel.kt
@@ -12,7 +12,7 @@ import hsk.practice.myvoca.data.toWordClass
 import hsk.practice.myvoca.module.LocalVocaPersistence
 import hsk.practice.myvoca.room.vocabulary.toVocabularyImplList
 import hsk.practice.myvoca.ui.state.UiState
-import hsk.practice.myvoca.xor
+import hsk.practice.myvoca.util.xor
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.MutableStateFlow
