+++ app/src/main/java/hsk/practice/myvoca/work/CreateTodayWordWork.kt
@@ -6,26 +6,25 @@ import androidx.work.*
 import com.orhanobut.logger.Logger
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
-import hsk.practice.myvoca.randoms
 import hsk.practice.myvoca.room.RoomVocaDatabase
 import hsk.practice.myvoca.room.todayword.RoomTodayWord
 import hsk.practice.myvoca.room.todayword.TodayWordDao
 import hsk.practice.myvoca.room.vocabulary.VocaDao
 import hsk.practice.myvoca.ui.screens.home.getSecondsLeft
+import hsk.practice.myvoca.util.randoms
+import hsk.practice.myvoca.util.writeLogToFile
 import kotlinx.coroutines.flow.first
 import java.util.concurrent.TimeUnit
-import javax.inject.Inject
 
 @HiltWorker
 class CreateTodayWordWorker @AssistedInject constructor(
     @Assisted context: Context,
-    @Assisted workerParams: WorkerParameters
+    @Assisted workerParams: WorkerParameters,
+    private val database: RoomVocaDatabase
 ) : CoroutineWorker(context, workerParams) {
 
     private val todayWordSize = 5
 
-    @Inject
-    lateinit var database: RoomVocaDatabase
     private val vocaDao: VocaDao
         get() = database.vocaDao()!!
     private val todayWordDao: TodayWordDao
@@ -38,6 +37,11 @@ class CreateTodayWordWorker @AssistedInject constructor(
             val todayWords = allVocabulary.map { it.id }.randoms(todayWordSize)
                 .map { RoomTodayWord(vocabularyId = it, checked = false) }
             todayWordDao.insertTodayWord(todayWords)
+            writeLogToFile(
+                context = applicationContext,
+                filename = "today-word-worker.txt",
+                log = "Save Today word - $todayWords"
+            )
             Result.success()
         } catch (e: Throwable) {
             Logger.e(e, "Error while refreshing Today's Word")
@@ -46,13 +50,13 @@ class CreateTodayWordWorker @AssistedInject constructor(
     }
 }
 
-val createTodayWordWorkerTag = "create_today_word_work"
+const val createTodayWordWorkerTag = "create_today_word_work"
 
 fun setPeriodicTodayWordWork(workManager: WorkManager) {
     val secondsLeft = getSecondsLeft()
     val periodicWork = PeriodicWorkRequestBuilder<CreateTodayWordWorker>(1, TimeUnit.DAYS)
         .addTag(createTodayWordWorkerTag)
-        .setInitialDelay(secondsLeft, TimeUnit.MILLISECONDS)
+        .setInitialDelay(secondsLeft, TimeUnit.SECONDS)
         .build()
     workManager.enqueueUniquePeriodicWork(
         createTodayWordWorkerTag,
@@ -61,6 +65,7 @@ fun setPeriodicTodayWordWork(workManager: WorkManager) {
     )
 }
 
+// TODO: 잘 안되면 workmanager 대신 application에서 처리해 버릴 수도 있음
 fun setOneTimeTodayWordWork(workManager: WorkManager) {
     val oneTimeWork = OneTimeWorkRequestBuilder<CreateTodayWordWorker>()
         .addTag(createTodayWordWorkerTag)
